+++ src/main/java/com/airline/api/repositories/AircraftRepository.java
@@ -1,9 +1,11 @@
-// This is the repository class for the Aircraft entity. This class is used to interact with the database.
-
 package com.airline.api.repositories;
 
 import com.airline.api.models.Aircraft;
+import com.airline.api.models.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AircraftRepository extends JpaRepository<Aircraft, Integer> {
+import java.util.List;
+
+public interface AircraftRepository extends JpaRepository<Aircraft, Long> {
+
 }
